upgraded protractor and reverted e2e test names
The latest version of protractor has a breaking change in where you
import the "globals" from.

Since the CLI tool generates a protractor config that expects e2e files
to be of the form *.e2e-spec.ts this commit reverts to that so that we
are consistent with the generated blueprint.